Use Forwarded and X-Forwarded-* headers when calculating server URL
When the application is served behind a proxy (e.g. for load balancing),
the proxy can provide the de-facto standard X-Forwarded-Proto, X-Forwarded-Host
and X-Forwarded-Port headers to submit the protocol, host and port of the original
request to the downstream application.

Using Spring's UriComponentsBuilder we can use this information to construct
the server URL of the original request.